+++ .isort.cfg
@@ -5,4 +5,4 @@ include_trailing_comma=True
 force_grid_wrap=0
 use_parentheses=True
 known_first_party = molecad
-known_third_party = fastapi,loguru,mongordkit,pathlib,pydantic,pymongo,pytest,rdkit,requests,starlette,starlette_context,typing_extensions,uvicorn
+known_third_party = fastapi,loguru,pathlib,pydantic,pymongo,pytest,rdkit,requests,starlette,starlette_context,typing_extensions,uvicorn
