+++ mypy.ini
@@ -1,14 +1,22 @@
 [mypy]
 python_version = 3.7
+plugins = pydantic.mypy
+
+follow_imports = silent
+warn_redundant_casts = True
+warn_unused_ignores = True
+disallow_any_generics = True
+check_untyped_defs = True
+no_implicit_reexport = True
+
+[pydantic-mypy]
+init_forbid_extra = True
+init_typed = True
+warn_required_dynamic_aliases = True
+warn_untyped_fields = True
 
 [mypy-pytest]
 ignore_missing_imports = True
 
-[mypy-pymongo.*]
-ignore_missing_imports = True
-
 [mypy-rdkit.*]
-ignore_missing_imports = True
-
-[mypy-pandas]
 ignore_missing_imports = True
