+++ old/api.py
@@ -4,7 +4,7 @@ from fastapi import FastAPI
 from pydantic import BaseModel, NonNegativeInt, PositiveInt
 
 from .api_db import compound_search, compound_search_summary
-from .settings import settings
+from src.settings import settings
 
 
 class Compound(BaseModel):
