+++ old/downloader_types.py
